requirements.txt
Flask==2.3.2
Flask-Dance==6.2.0
Flask-SQLAlchemy==3.0.3
qrcode==7.4.2
Pillow==9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
appwrite==2.0.0
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0
